reference_bot/requirements.txt
aiofiles
aiogram>=3.11.0
aiogram.contrib
aiogram_dialog>=2.2.0
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
APScheduler>=3.11.0
asyncpg>=0.30.0
attrs
cachetools
certifi
charset-normalizer
docker
environs
frozenlist
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
greenlet
gspread
httplib2
idna
Jinja2
magic-filter
MarkupSafe
marshmallow
multidict
oauthlib
propcache
proto-plus
protobuf
psycopg>=3.2.9
psycopg-pool>=3.2.6
psycopg2-binary
pyasn1
pyasn1_modules
pydantic>=2.11.0
pydantic_core
pyparsing
python-dotenv
pytz
redis>=5.0.0
requests
requests-oauthlib
rsa
six
sniffio
SQLAlchemy>=2.0.0
tenacity
typing-inspection
typing_extensions
tzlocal
Unidecode
uritemplate
urllib3
yarl